requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
pyjwt[crypto]==2.9.0
httpx==0.27.2
motor==3.6.0
cryptography==43.0.3
python-multipart==0.0.12
